Time: 0 ms (100.00%) | Memory: 113.5 MB (61.53%) - LeetSync

+++ 643-maximum-average-subarray-i/maximum-average-subarray-i.cpp
@@ -14,6 +14,6 @@ public:
             max_sum = max(max_sum, curr_sum);
         }
 
-        return static_cast<double>(max_sum) / k;
+        return static_cast<double>(max_sum) / double(k);
     }
 };
